requirements.txt
PyQt6
rembg[cpu]
Pillow
onnxruntime
